Update layout to reuse in EE
