+++ ui/index.ts
@@ -114,6 +114,8 @@ import {
   statusSortHelper,
 } from "./lib/utils";
 import SignIn from "./pages/SignIn";
+import Layout from "./components/Layout";
+import useNavigation from "./hooks/navigation";
 
 export {
   AppContext,
@@ -176,6 +178,7 @@ export {
   KubeStatusIndicator,
   KustomizationDetail,
   Kustomization,
+  Layout,
   Link,
   LinkResolverProvider,
   LoadingPage,
@@ -228,6 +231,7 @@ export {
   useSyncFluxObject,
   useRequestState,
   useToggleSuspend,
+  useNavigation,
   UserSettings,
   V2Routes,
   YamlView,
